refactor(93): Refactor TerminalUtils, Add more logs to BuildExecutionService
